Safari-609.1.11.1/LayoutTests/editing/execCommand/delete-line-and-insert-text-in-font-inside-blockquote.html

<!DOCTYPE html>
<html>
<body>
<div id="container" contenteditable>
<blockquote style="background-color:yellow"><br><font style="color:blue"><br>&nbsp;world
</font></blockquote>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests deleting a line inside a font element inside a blockquote.')

var font = document.querySelector('font');
getSelection().collapse(font, 0);
getSelection().modify('Extend', 'Forward', 'Line');
document.execCommand('Delete', false, null);
document.execCommand('InsertText', false, 'hello');

Markup.dump('container', 'There should be exactly one blockquote and you should see "hello world" below')

</script>
</body>
</html>
