third_party/blink/web_tests/platform/generic/editing/execCommand/delete-line-and-insert-text-in-font-inside-blockquote-expected.txt

This tests deleting a line inside a font element inside a blockquote.

There should be exactly one blockquote and you should see "hello world" below:
| "
"
| <blockquote>
|   style="background-color:yellow"
|   <br>
|   <font>
|     style="color:blue"
|     "hello<#selection-caret> world
"
| "
"
